+++ Sources/Views/InputStackView.swift
@@ -69,12 +69,4 @@ open class InputStackView: UIStackView {
         distribution = .fill
         alignment = .bottom
     }
-    
-    open override func layoutIfNeeded() {
-        super.layoutIfNeeded()
-        
-        // We need to invalidate the size of the superview (MessageInputBar) in case the subview heights have changed
-        // and thus the superview size needs to change
-        superview?.invalidateIntrinsicContentSize()
-    }
 }